requirements.txt
Flask==2.3.2
Flask-WTF==1.1.1
python-dotenv==1.0.0
psycopg2-binary==2.9.6
bcrypt==4.0.1
requests==2.31.0